+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Developer changes:
 
 Beta 0.4.3
 ----------
+Added functionality:
+* `directory` parsers now support full file paths for upload.
+
 Developer changes:
 * Added method `get_metadata(self, sample_id, project_id)` to allow other metadata to be fetched from the `api` module
 * Added method `send_metadata(self, metadata, project_id, sample_id)` to allow other metadata to be sent using the `api` module